Addition of FM demod
Two additional demonstration applications were added,illustrating the demodulation of speech FM (frequency modulation, the modulation of F0), i.e. the frequency patterns of tones, pitch accents and intonations. The first is first for demonstrating the steps in AMDF (Average Magnitude Difference Function) demodulation, the second is for visualising the output of AMDF FM demodulation.